Removed SelectATermScreen class, renamed the file and removed Select a term from navigation
